% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(trials = 1, success = 1)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{success}{a non-negative integer, smaller or equal to n, which provides the number of successes in n trials}
}
\value{
an object of numeric class, number of permutations for given success and trials
}
\description{
function which computes number of permutations for k successes in n trials
}
\examples{
#default
1_choose_1 <- bin_choose(1, 1)

#second example
10_choose_4 <- bin_choose(10, 4)

#0 successes
5_choose_0 <- bin_choose(5, 0)
}
