% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of bincum class, cumulative probability of binomial distribution
}
\description{
function which computes culmulative probablity for binomial distributions for given inputs trials and prob
}
\examples{

#default
bin_cumulative(1, 0.5)

#example 2
bin_cumulative(10, 0.8)

}
