% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of numeric class, kurtosis of binomial distribution
}
\description{
function which computes kurtosis of binomial distribution with given inputs trials and prob
}
\examples{

#default
binkurtosis1 <- bin_kurtosis(10, 0.3)

#example2
binkurtosis2 <- bin_kurtosis(100, 0.2)

}
