% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of numeric class, mode of binomial distribution
}
\description{
function which computes mode of binomial distribution with given inputs trials and prob
}
\examples{

#default
binmode1 <- bin_mean(10, 0.3)

#example2
binmode2 <- bin_mean(5, 0.5)

}
