% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(trials, success, prob)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{success}{a non-negative integer, smaller or equal to n, which provides the number of successes in n trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of numeric class, binomial probability for given successes and trials
}
\description{
function which computes binomial probability for k successes in n trials
}
\examples{

#default
prob1 <- bin_probability(1, 1, 0.5)

#multiple number of successes
prob2 <- bin_probablity(10, 1:5, 0.3)

}
