% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of numeric class, skewness of binomial distribution
}
\description{
function which computes skewness of binomial distribution with given inputs trials and prob
}
\examples{

#default
binskewness1 <- bin_skewness(10, 0.3)

#example2
binskweness2 <- bin_skewness(100, 0.2)

}
