% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of numeric class, variance of binomial distribution
}
\description{
function which computes variance of binomial distribution with given inputs trials and prob
}
\examples{

#default
binvar1 <- bin_variance(10, 0.3)

#example2
binvar2 <- bin_variance(100, 0.2)

}
