% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{probability}
\alias{probability}
\title{bin_distribution}
\format{An object of class \code{numeric} of length 1.}
\usage{
probability
}
\arguments{
\item{trials}{a non-negative integer which provides the number of trials}

\item{prob}{a number between 0 and 1 which gives the probablity of a success on an individual draw}
}
\value{
an object of bindis class, binomial distribution of probabilities
}
\description{
function which computes probabilities for getting exactly k successes in n trials with probability p
}
\examples{

default
dis1 <- bin_distribution(1, 0.5)

#example 2
dis2 <- bin_distribution(10, 0.2)


}
\keyword{datasets}
