% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the number of items to choose from (numeric)}

\item{k}{the number of items to choose (numeric)}
}
\value{
the number of combinations of k items (numeric)
}
\description{
finds the number of ways to choose k items from n
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
