% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success}
}
\value{
the number of combinations in which k successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
