% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
an object containing the cumulative distribution with the given parameters (bincum, data.frame)
}
\description{
creates a bincum and data.frame object
}
\examples{
dis2 <- bin_cumulative(trials = 5, prob = 0.5)
plot(dis2)
}
