% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mode}
\alias{bin_mode}
\title{binomial mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
the mode of the binomial distribution (numeric)
}
\description{
finds the mode of the binomial distribution
}
\examples{
bin_mode(10, 0.3)
}
