% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes (numeric)}

\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of a success (numeric)}
}
\value{
the probability of success successes (numeric)
}
\description{
finds the probability of the given number of successes with the given parameters
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
