% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{bin_probability}
\alias{bin_probability}
\title{Binomial Probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{Number of successes (numeric)}

\item{trials}{Number of trials  (numeric)}

\item{prob}{Probability of success in one trial (numeric)}
}
\value{
Returns the probability given certain number of successes, probability and trials
}
\description{
Calculates probability given certain number of successes, probability and trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
