% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the number of trials (numeric)}

\item{prob}{the probability of one success (numeric)}
}
\value{
the variance of the binomial distribution (numeric)
}
\description{
finds the variance of the binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
